Solve ohci-hcd problems of AMD IWBs
This commit solves ohci-hcd problems of AMD Interactive White Boards.
Driver works fine with ehci and xhci, but it fails to keep track of
touching fingers when it comes to ohci. This commit solves this problem
implementing a circular queue. Before this commit, the driver would get
64 bytes from usb_interrupt_msg and server daemon would get this 64-byte
chunk through the read system call. By the help of a circular queue,
the driver gets multiple of 64 bytes and sends a single 64-byte chunk
to the server when it calls read system call.

Fix indentation style and format.

Update README.md.

Update ChangeLog.